requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.3.0
psycopg2==2.6.1
drf-nested-routers==0.10.0
django-multiselectfield==0.1.3
